# ec_backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/query": {
      "get": {
        "operationId": "UserController_fetchUserByEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "query",
            "schema": {
              "title": "Email",
              "example": "john.doe@mail.com",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The user has been successfully fetched!"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "The email field must be a valid email (example: user@mail.com)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The user was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UserController_fetchUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The user has been successfully fetched!"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "The email field must be a valid email (example: user@mail.com)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The user was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The user has been successfully updated."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (uuid is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The user was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The user has been successfully deleted."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (uuid is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The user was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/notifications/{id}": {
      "get": {
        "operationId": "NotificationController_fetchNotificationById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The notification has been successfully fetched."
              }
            }
          },
          "400": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "Validation failed (numeric string is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The notification was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      },
      "patch": {
        "operationId": "NotificationController_updateNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The notification has been successfully updated."
              }
            }
          },
          "400": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "Validation failed (numeric string is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The notification was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      },
      "delete": {
        "operationId": "NotificationController_deleteNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The notification has been deleted."
              }
            }
          },
          "400": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "Validation failed (numeric string is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The notification was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/users/notifications": {
      "post": {
        "operationId": "UserNotificationController_generateUserNotification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/generateUserNotificationDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "A new notification has been generated!"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "The userId field must be a UUID."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "UserNotifications"
        ]
      }
    },
    "/api/users/notifications/query": {
      "get": {
        "operationId": "UserNotificationController_fetchUserNotifications",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The notifications have been successfully fetched."
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "There are no notifications to show."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "UserNotifications"
        ]
      }
    },
    "/api/users/notifications/update": {
      "patch": {
        "operationId": "UserNotificationController_updateUserNotification",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserNotificationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The notification has been updated!"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "Please, check the request. Either the user or the desired notification do not exist"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "UserNotifications"
        ]
      }
    },
    "/api/users/notifications/{id}": {
      "delete": {
        "operationId": "UserNotificationController_deleteUserNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The notification has been deleted."
              }
            }
          },
          "400": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "Validation failed (uuid is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The notification was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "UserNotifications"
        ]
      }
    },
...
    "/api/videos/progress/user/{userId}": {
      "get": {
        "operationId": "VideoProgressController_fetchVideoProgressesByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The progress connections have been successfully fetched!"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (uuid is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "There are no progress connections to show or the userId is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "VideoProgress"
        ]
      }
    },
...
    "/api/logs/user/{userId}": {
      "get": {
        "operationId": "LoginLogController_fetchUserLogs",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "Logs successfully fetched."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (uuid is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "There are no logs to show or the userId is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "LoginLog"
        ]
      }
    },
...
    "/api/slideshow/progress/user/{userId}": {
      "get": {
        "operationId": "SlideshowProgressController_fetchProgressesByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The progresses have been fetched."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (uuid is expected)"
              }
            }
          },
          "404": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "There are no progresses to show."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "SlideshowProgress"
        ]
      }
    },
...
  },
  "info": {
    "title": "English Conversations API",
    "description": "REST API for managing English Conversations' data and processes.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpdateNotificationDTO": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "example": "INFO"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "example": "Congratulations!"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "example": "You have just completed chapter A1!"
          },
          "actionUrl": {
            "type": "string",
            "title": "ActionURL",
            "example": "https://google.com"
          },
          "expirationDate": {
            "type": "string",
            "title": "ExpirationDate",
            "example": "2025-07-16"
          }
        },
        "required": [
          "type",
          "title"
        ]
      },
      "generateUserNotificationDTO": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "title": "UserID",
            "example": "$F0g9219*"
          },
          "notificationId": {
            "type": "string",
            "title": "NotificationID",
            "example": "10854fkfa8**93mNN"
          }
        },
        "required": [
          "userId",
          "notificationId"
        ]
      },
      "UpdateUserNotificationDTO": {
        "type": "object",
        "properties": {
          "isRead": {
            "type": "boolean",
            "title": "IsRead",
            "example": true
          },
          "readAt": {
            "type": "string",
            "title": "ReadAt",
            "example": "2025-10-10"
          },
          "deliveredViaEmail": {
            "type": "boolean",
            "title": "DeliveredViaEmail",
            "example": true
          },
          "deliveredViaApp": {
            "type": "boolean",
            "title": "DeliveredViaApp",
            "example": false
          }
        }
      },
...
    }
  }
}